potato_leaf_hybrid_project/requirements.txt
numpy
tensorflow
pennylane
scikit-learn
matplotlib
pandas
pillow
streamlit
